requirements.txt
streamlit
pandas
altair
python-docx
pypdf
docxtpl
reportlab
google-generativeai